Рефакторинг: 1. добавлены комментарии к контроллерам и к swagger; 2. добавлены CancellationToken в асинхронные методы; 3. добавлена валидация данных в бизнес-логику. Исправления: 1. при удалении поста/тега теперь корректно удаляются связи из промежуточной таблицы PostTags; 2. при сохранении поста теперь корректно обновляются связи Many-to-Many с тегами.

+++ BackendService.BLL/Logics/PostLogic.cs
@@ -1,4 +1,5 @@
-﻿using BackendService.BLL.Interfaces;
+﻿using BackendService.Common.Exceptions;
+using BackendService.BLL.Interfaces;
 using BackendService.Common.DTO;
 
 namespace BackendService.BLL.Logics
@@ -14,17 +15,39 @@ namespace BackendService.BLL.Logics
 
         public async Task<PostDTO?> GetPostById(int id, CancellationToken token = default)
         {
-            return await _postRepository.GetPostById(id, token);
+            if (id <= 0) throw new ValidationException("ID должен быть положительным целым числом");
+
+            var post = await _postRepository.GetPostById(id, token);
+
+            return post is null ? throw new NotFoundException($"Пост с ID {id} не найден") : post;
         }
 
         public async Task DeletePost(int id, CancellationToken token = default)
         {
-            await _postRepository.DeletePost(id, token);
+            if (id <= 0) throw new ValidationException("ID должен быть положительным целым числом");
+
+            try
+            {
+                await _postRepository.DeletePost(id, token);
+            }
+            catch (InvalidOperationException)
+            {
+                throw new NotFoundException($"Пост с ID {id} не найден и не может быть удалён");
+            }
         }
 
         public async Task<PostEditDTO> SavePost(PostEditDTO post, CancellationToken token = default)
         {
-            return await _postRepository.SavePost(post, token);
+            if (post.Id < 0) throw new ValidationException("ID должен быть положительным целым числом");
+
+            try
+            {
+                return await _postRepository.SavePost(post, token);
+            }
+            catch (InvalidOperationException)
+            {
+                throw new NotFoundException($"Пост с ID {post.Id} не найден и не может быть отредактирован");
+            }
         }
     }
 }
